Validation and some Tests added

+++ backend/src/controllers/IncidentController.js
@@ -5,15 +5,29 @@ module.exports = {
     async delete(request,response){
         const { id } = request.params;
         const ong_id = request.headers.authorization;
-
+         
         const incident = await connection('incidents')
             .where('id', id)
             .select('ong_id')
             .first();
-
+        
+        /* 
+        EN:
+            EDIT SUMMARY:
+                #Added to check if the user given incident id correlates to a entry in the database averting a unhandled promise rejection
+        PT-BR:    
+            RESUMO DA EDIÇÃO:
+                #Adicionado para verificiar se o id de incident dado pelo usuário foi encontrado no banco de dados evitando assim uma rejeição de promessa sem resposta
+        */    
+        if(!incident){
+            return response.status(406).json({error: "Not Acceptable, Incident not found"});
+        }
+        
         if(incident.ong_id !== ong_id){
             return response.status(401).json({error: "Operation not permitted."});
         }
+       
+       
 
         await connection('incidents').where('id', id).delete();
